+++ lib/chef/scan_access_control/windows.rb
@@ -1,0 +1,41 @@
+#--
+# Author:: Tyler Cloke (<tyler@opscode.com>)
+# Copyright:: Copyright (c) 2012 Opscode, Inc.
+# License:: Apache License, Version 2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+class Chef
+  class ScanAccessControl
+    class Windows
+      
+      # Modifies @current_resource, setting the current ACL state.
+      def set_all!
+        if ::File.exist?(new_resource.path)
+          # for each person, set sid
+            # for each sid, set {type, access, inherited_from}
+          puts "new resource"*100
+          puts @new_resource
+          puts "current resource"*100
+          puts @current_resouce          
+        else
+          # leave the values as nil.
+        end
+      end
+      
+      
+      
+    end
+  end
+end
